Add new font and functionality to menu screens
