+++ Assets/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 838ddf838ac5aaf4e9863a28c5e49351
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
