--- MainMenu.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class MainMenu : MonoBehaviour
-{
-    public void ExitButton()
-    {
-        Application.Quit();
-        Debug.Log("Game Closing");
-    }
-
-    public void StartGame()
-    {
-        SceneManager.LoadScene("Game");
-    }
-
-}
