--- PauseMenu.cs
@@ -1,52 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class PauseMenu : MonoBehaviour
-{
-    public static bool GameIsPaused = false;
-
-    public GameObject pauseMenuUI;
-
-    void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Escape))
-        {
-            if (GameIsPaused)
-            {
-                ResumeButton();
-            }
-            else
-            {
-                PauseButton();
-            }
-        }
-    }
-
-    public void ResumeButton()
-    {
-        pauseMenuUI.SetActive(false);
-        Time.timeScale = 1f;
-        GameIsPaused = false;
-    }
-    void PauseButton()
-    {
-        pauseMenuUI.SetActive(true);
-        Time.timeScale = 0f;
-        GameIsPaused = true;
-    }
-
-    public void MenuButton()
-
-    {
-        Time.timeScale = 1f;
-        SceneManager.LoadScene("Menu");
-    }
-
-    public void QuitButton()
-    {
-        Debug.Log("Quitting game...");
-        Application.Quit();
-    }
-}
