+++ ProjectSettings/EditorBuildSettings.asset
@@ -4,5 +4,17 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-  m_Scenes: []
+  m_Scenes:
+  - enabled: 1
+    path: Assets/Scenes/TDMainMenu.unity
+    guid: 5eb9a912b31d1e347bff66043c5186f0
+  - enabled: 1
+    path: Assets/Scenes/GameOverScreen.unity
+    guid: 41a761c3faf6dc349854b187668b6ac4
+  - enabled: 1
+    path: Assets/Scenes/PauseMenu.unity
+    guid: 90c9a052413183245aa1ff69ff6a334a
+  - enabled: 1
+    path: Assets/Scenes/SampleScene.unity
+    guid: 2cda990e2423bbf4892e6590ba056729
   m_configObjects: {}
